+++ Sources/MarmitonUIMealzIOS/ui/RecipeDetails/MarmitonRecipeDetailsTagsView.swift
@@ -1,0 +1,17 @@
+//
+//  MarmitonRecipeDetailsTagsView.swift
+//
+//
+//  Created by didi on 22/10/2024.
+//
+
+import SwiftUI
+import MealziOSSDK
+
+@available(iOS 14, *)
+public struct MarmitonRecipeDetailsTagsView: RecipeDetailsTagsProtocol {
+    public init() {}
+    public func content(params: RecipeDetailsTagsParameters) -> some View {
+       EmptyView()
+    }
+}